+++ java-vtl-script/src/main/java/no/ssb/vtl/script/visitors/join/FoldVisitor.java
@@ -21,7 +21,6 @@ package no.ssb.vtl.script.visitors.join;
  */
 
 import com.google.common.collect.ImmutableSet;
-import no.ssb.vtl.model.Component;
 import no.ssb.vtl.model.Dataset;
 import no.ssb.vtl.parser.VTLParser;
 import no.ssb.vtl.script.operations.FoldOperation;
@@ -47,8 +46,9 @@ public class FoldVisitor extends VTLDatasetExpressionVisitor<FoldOperation> {
         String dimension = ctx.dimension.getText();
         String measure = ctx.measure.getText();
 
-        Set<Component> elements = ctx.variableExpression().stream()
+        Set<String> elements = ctx.variableExpression().stream()
                 .map(componentVisitor::visit)
+                .map(dataset.getDataStructure()::getName)
                 .collect(ImmutableSet.toImmutableSet());
 
         return new FoldOperation(dataset, dimension, measure, elements);