+++ CHANGELOG.md
@@ -1,10 +1,25 @@
 # Changelog 
 
+## develop
+
+### Added
+
+### Changed
+
+## 0.1.10 - 2018-08-31
+
+### Added
+
+* Support for average aggregation function `ds := avg(ds1) group by ds1.x`
+* Attribute components are now kept when using `fold`
+* Fold optimization
+* Expose keyword list in `VTLScriptEngine`
+
 ## 0.1.9-2 - 2018-06-01
 
 ### Changed
 
-- Force type casting to ensure correct return type when VTLFloor gets input of type Integer.
+* Force type casting to ensure correct return type when VTLFloor gets input of type Integer.
 
 ## 0.1.9-1 - 2018-05-30